+++ api/endpoint_LEGACY/hooks/projects.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { getMemberProfileById } from '@/api/legacy/members';
-import { getProjectById, getProjects } from '@/api/legacy/projects';
+import { getMemberProfileById } from '@/api/endpoint_LEGACY/members';
+import { getProjectById, getProjects } from '@/api/endpoint_LEGACY/projects';
 
 // project id로 조회
 export const useGetProjectById = (id?: string) => {
