+++ api/endpoint_LEGACY/image/index.ts
