+++ components/common/Header/index.tsx
@@ -2,7 +2,7 @@ import Link from 'next/link';
 import { useRouter } from 'next/router';
 import { FC } from 'react';
 
-import { useGetMemberOfMe } from '@/api/legacy/hooks';
+import { useGetMemberOfMe } from '@/api/endpoint_LEGACY/hooks';
 import useAuth from '@/components/auth/useAuth';
 import DesktopHeader from '@/components/common/Header/desktop/DesktopHeader';
 import MobileHeader from '@/components/common/Header/mobile/MobileHeader';
