+++ components/members/detail/CareerSection/index.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import Link from 'next/link';
 import LinkIcon from 'public/icons/icon-link.svg';
 
-import { MemberLink } from '@/api/legacy/members/type';
+import { MemberLink } from '@/api/endpoint_LEGACY/members/type';
 import CareerItem from '@/components/members/detail/CareerSection/CareerItem';
 import InfoItem from '@/components/members/detail/InfoItem';
 import MemberDetailSection from '@/components/members/detail/MemberDetailSection';
