+++ components/members/detail/MemberDetail.tsx
@@ -9,8 +9,8 @@ import MailIcon from 'public/icons/icon-mail.svg';
 import ProfileIcon from 'public/icons/icon-profile.svg';
 import { FC, useMemo } from 'react';
 
-import { useGetMemberProfileById } from '@/api/legacy/hooks';
-import { isProjectCategory } from '@/api/legacy/projects/type';
+import { useGetMemberProfileById } from '@/api/endpoint_LEGACY/hooks';
+import { isProjectCategory } from '@/api/endpoint_LEGACY/projects/type';
 import Loading from '@/components/common/Loading';
 import useEventLogger from '@/components/eventLogger/hooks/useEventLogger';
 import CareerSection from '@/components/members/detail/CareerSection';
