+++ components/members/detail/MessageSection/MessageModal.tsx
@@ -6,7 +6,7 @@ import { FC, useState } from 'react';
 import { useForm } from 'react-hook-form';
 import * as yup from 'yup';
 
-import { usePostCoffeeChatMutation } from '@/api/legacy/hooks';
+import { usePostCoffeeChatMutation } from '@/api/endpoint_LEGACY/hooks';
 import RHFControllerFormItem from '@/components/common/form/RHFControllerFormItem';
 import Input from '@/components/common/Input';
 import Loading from '@/components/common/Loading';
