+++ components/projects/main/ProjectCard.tsx
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import Link from 'next/link';
 import { FC } from 'react';
 
-import { ProjectDetail } from '@/api/legacy/projects/type';
+import { ProjectDetail } from '@/api/endpoint_LEGACY/projects/type';
 import ResizedImage from '@/components/common/ResizedImage';
 import Text from '@/components/common/Text';
 import useEventLogger from '@/components/eventLogger/hooks/useEventLogger';
