+++ components/projects/main/ProjectDetail.tsx
@@ -4,8 +4,8 @@ import Link from 'next/link';
 import { useRouter } from 'next/router';
 import { FC, useMemo, useState } from 'react';
 
-import { useGetMemberOfMe } from '@/api/legacy/hooks';
-import { deleteProject } from '@/api/legacy/projects';
+import { useGetMemberOfMe } from '@/api/endpoint_LEGACY/hooks';
+import { deleteProject } from '@/api/endpoint_LEGACY/projects';
 import ConfirmModal from '@/components/common/Modal/Confirm';
 import MemberBlock from '@/components/members/common/MemberBlock';
 import WithMemberMetadata from '@/components/members/common/WithMemberMetadata';
