+++ components/projects/upload/hooks/useGetMembersByNameQuery.ts
@@ -1,6 +1,6 @@
 import { useQuery } from '@tanstack/react-query';
 
-import { getMembersSearchByName } from '@/api/legacy/members';
+import { getMembersSearchByName } from '@/api/endpoint_LEGACY/members';
 interface GetMembersByNameQueryVariables {
   name: string;
 }
